// src/setup/puppeteer.ts
import { DappeteerBrowser } from "../browser";
import { DappeteerLaunchOptions } from "../types";

export async function launchPuppeteer(
  metamaskPath: string,
  userDataDir: string,
  options: DappeteerLaunchOptions
): Promise<DappeteerBrowser> {
  const useExtra = options.usePuppeteerExtra ?? false;
  const launchOpt = {
    ...(options.puppeteerOptions ?? {}),
    headless: options.headless,
    userDataDir,
    args: [
      "--accept-lang=en",
      "--window-size=1920,1080",
      `--disable-extensions-except=${metamaskPath}`,
      `--load-extension=${metamaskPath}`,
      ...(options.puppeteerOptions?.args || []),
      ...(options.headless ? ["--headless=new"] : []),
    ],
  };

  if (useExtra) {
    console.info("goint to using puppeteer-extra");
  }
  const pBrowser = useExtra
    ? await (await import("puppeteer-extra")).default.launch(launchOpt)
    : await (await import("puppeteer")).default.launch(launchOpt);

  const { DPuppeteerBrowser } = await import("../puppeteer");
  return new DPuppeteerBrowser(pBrowser, userDataDir, options.metaMaskFlask);
}
